Wire up concept route with search modal
