+++ arches_lingo/urls.py
@@ -11,6 +11,7 @@ urlpatterns = [
     path("login", LingoRootView.as_view(), name="login"),
     path("advanced-search", LingoRootView.as_view(), name="advanced-search"),
     path("schemes", LingoRootView.as_view(), name="schemes"),
+    path("concept/<uuid:id>", LingoRootView.as_view(), name="concept"),
     path("api/concepts", ConceptTreeView.as_view(), name="api-concepts"),
     path("api/search", ValueSearchView.as_view(), name="api-search"),
     path("", include("arches_references.urls")),
